pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wraval"
version = "0.1.0"
description = """WRAVAL helps in evaluating LLMs for writing assistant tasks like summarization,
professional tone, witty tone, etc."""
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "Gabriel Benedict", email = "gbndict@amazon.com" }]

# Dependencies – see note below for loading from requirements.txt
dependencies = [
    "tqdm~=4.66.4",
    "pandas~=2.2.3",
    "beautifulsoup4~=4.12.3",
    "boto3",
    "plotly~=5.24.1",
    "transformers==4.51.0",
    "datasets~=3.2.0",
    "evaluate~=0.4.3",
    "dynaconf~=3.2.7",
    "torch",
    "botocore",
    "sagemaker==2.244.2",
    "numpy",
    "requests",
    "accelerate",
    "torchvision",
    "typer",
    "dspy-ai>=2.5.0"
]

[project.scripts]
wraval = "wraval.main:main"

[project.optional-dependencies]
gpu = ["bitsandbytes==0.45.5"]

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.data-files]
# This copies the config file into the installation (non-package dir)
"config" = ["config/settings.toml"]